requirements.txt
opencv-python>=4.10.0.84
numpy>=2.2.6
Pillow>=12.0.0
pillow-heif>=1.1.1
paddleocr>=3.3.1
pytesseract>=0.3.13
easyocr>=1.7.2
pandas>=1.3.0
requests>=2.25.0
python-dotenv>=0.19.0
slack-sdk>=3.19.0